copilot_chat: Cleanup after removing oauth via LSP
