+++ crates/copilot_chat/Cargo.toml
@@ -22,16 +22,19 @@ test-support = [
 [dependencies]
 anthropic.workspace = true
 anyhow.workspace = true
+collections.workspace = true
 credentials_provider.workspace = true
 dirs.workspace = true
 futures.workspace = true
 gpui.workspace = true
 http_client.workspace = true
+language_model.workspace = true
 log.workspace = true
 paths.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 settings.workspace = true
+util.workspace = true
 zed_credentials_provider.workspace = true
 
 [dev-dependencies]
